[#717] bug fix unknown property in yaml test
